Add io_ops.hh
Defines templated IO functions in the scolex::io namespace. These are
more or less the same as q_* functions, albeit with no expectation of
there being an abstract base class to use.

Upside: no need for virtual functions. It's simply a compiler error to
do something with a stream that isn't compatible, except in the case of
io::eof, which returns false if unimplemented.

Downside: potential code bloat, though that situation's improved a lot
so I'm not that concerned about it. Other downside: probably hard to
export.